Updates TableD with latest guidance
This commit updates TableD with the latest changes for animal tagged
data. Thsi commit also adds a utility function to retrieve a particular
code figure for a code table. This commit also improves the performance
of looking up code figures by using the csv module instead of pandas
which is a very heavyweight approach.