Refactors encoding functions into their own module
The primary encoding functions for building the BUFR message existed in
the `encode_animal_tag` module but since they are required for every
type of BUFR encoding and should not be coupled to the type of encoding
that's happening they should be moved into their own module. This commit
moves the general encoding functions into the `bufrtools.encoding.bufr`
module.