Replace netCDF as the base data source with pandas DataFrames
To allow for in-pipeline processing without converting to a netCDF file
first I made changes to be able to pass in a path to a CSV or Parquet
file in addition to a netCDF file.

The CSV and Parquet files have no metadata so the uuid and ptt attributes
were added as parameters. If a netCDF file is the data source those
attributes are extracted from the file and the parameters are ignored.

Also removed an activate script from root dir