Fixes unit tests
A while back when we made the change of units from deg_C to Kelvin for
temperature values, the temperature field overlapped with the same octet
that described the replication factor (i.e. number of profiles to
follow) which broke the tests. This commit updates the expected value to
match what it should be. In the MR I will include screenshots of the
hexdump for confirmation.